server/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-cors-headers>=4.4
python-dotenv>=1.0
psycopg2-binary>=2.9
celery>=5.4
redis>=5.0
django-storages>=1.14
boto3>=1.34
pytest>=8.0
pytest-django>=4.8
# Parsing PDF schedules
PyPDF2>=3.0
# Web push notifications
pywebpush>=1.14
pytz>=2024.1
# Opcionales:
# django-channels>=4.1
# channels-redis>=4.2
# weasyprint>=62.0
# beautifulsoup4>=4.12
